+++ artf/src/main/java/com/prjgrp/artf/repository/PaymentRepo.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.prjgrp.artf.model.Payment;
 
 @Repository
-public interface PaymentRepo extends JpaRepository<Payment, Integer> {
+public interface PaymentRepo extends JpaRepository<Payment, Long> {
 
 }
