+++ artf/src/main/java/com/prjgrp/artf/repository/PropertyRepo.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.prjgrp.artf.model.Property;
 
 @Repository
-public interface PropertyRepo extends JpaRepository<Property, Integer> {
+public interface PropertyRepo extends JpaRepository<Property, Long> {
 
 }
